requirements.txt
# Core libraries
numpy==1.19.5
scipy==1.5.4
scikit-learn==1.0.2
pillow==8.4.0
opencv-python==4.12.0.88
pandas==1.3.5

# Deep learning
tensorflow==2.6.0
keras==2.6.0
keras-applications==1.0.8
keras-preprocessing==1.1.2
git+https://github.com/rcmalli/keras-vggface.git

# Face detection
mtcnn==0.1.0

# Web app front-end
streamlit==1.2.0
altair==4.2.2
typing-extensions==4.13.2

# Utilities
h5py==3.1.0
protobuf==3.20.3


